Refactor nextjs app.
