+++ components/cards/avg-age/avg-age-card.jsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import { Card } from '../../layout/card/card';
-import './avg-age-card.module.css';
+import styles from './avg-age-card.module.css';
 import { formatDate } from '../../../utils/date';
+import { parseAverageAge } from '../../../utils/parse';
 
 export const EMBED_PATH_AVERAGE_AGE = 'varsta-medie';
 export class AverageAgeCard extends React.PureComponent {
   render() {
     const { state, title } = this.props;
-    const { isLoaded, error, lastUpdatedOn, averageAge, stale } = state;
+    const { error, lastUpdatedOn, averageAge, stale } = parseAverageAge(state);
     return (
       <Card
-        isLoaded={isLoaded}
         error={error}
         title={title}
         subtitle={`Ultima actualizare: ${formatDate(lastUpdatedOn)}`}
         isStale={stale}
         embedPath={EMBED_PATH_AVERAGE_AGE}
       >
-        <div className="circle">
-          <span className="value">{averageAge}</span>
+        <div className={styles.circle}>
+          <span className={styles.value}>{averageAge}</span>
         </div>
       </Card>
     );
