+++ components/cards/counties-table/counties-table.jsx
@@ -1,30 +1,18 @@
 import React from 'react';
 import Image from 'next/image'
-import './counties-table.module.css';
+import styles from './counties-table.module.css';
 import { Card } from '../../layout/card/card';
 import { formatDate } from '../../../utils/date';
+import { parseCountiesTable } from '../../../utils/parse';
 
 export const EMBED_COUNTIES_TABLE = 'counties-table';
 
 export class CountiesTable extends React.PureComponent {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      page: 0,
-      limit: 10,
-      isLoaded: false,
-      error: null,
-      counties: [],
-    };
-  }
-
-  componentWillReceiveProps(newProps) {
-    if (this.props !== newProps) {
-      const { state } = newProps;
-      this.setState({ ...state });
-    }
-  }
+  state = {
+    page: 0,
+    limit: 10,
+    parsedData: parseCountiesTable(this.props.state)
+  };
 
   displayTable(counties) {
     const { page, limit } = this.state;
@@ -34,7 +22,7 @@ export class CountiesTable extends React.PureComponent {
       .map((row, index) => {
         return (
           <tr key={`dailyTable${index}`}>
-            <td>{row.name}</td>
+            <td className={styles.td}>{row.name}</td>
             <td className="has-text-right">{row.countyInfectionsNumbers}</td>
             <td className="has-text-right">{row.value}</td>
           </tr>
@@ -43,22 +31,18 @@ export class CountiesTable extends React.PureComponent {
   }
 
   displayPagination() {
-    if (!this.state.counties) {
-      return null;
-    }
-
-    const { counties, limit, page } = this.state;
+    const { limit, page, parsedData: {counties} } = this.state;
     const shouldDisplayPagination = counties.length > limit;
     if (shouldDisplayPagination) {
       return (
-        <div className="navigation">
+        <div className={styles.navigation}>
           <div
             className={'button ' + (page === 0 ? 'hide' : '')}
             onClick={(e) => this.changePage(-1)}
           >
             <Image
               src="/images/chevrons-left.svg"
-              className="navigation-chevron"
+              className={styles.navigation_chevron}
               alt="Pagina anterioară"
               width={24}
               height={24}
@@ -73,7 +57,7 @@ export class CountiesTable extends React.PureComponent {
           >
             <Image
               src="/images/chevrons-right.svg"
-              className="navigation-chevron"
+              className={styles.navigation_chevron}
               alt="Pagina următoare"
               width={24}
               height={24}
@@ -85,7 +69,7 @@ export class CountiesTable extends React.PureComponent {
   }
 
   changePage(inc) {
-    const { page, counties, limit } = this.state;
+    const { page, limit, parsedData: {counties} } = this.state;
 
     if (inc < 0 && page !== 0) {
       this.setState({ page: page + inc });
@@ -97,23 +81,22 @@ export class CountiesTable extends React.PureComponent {
   }
 
   render() {
-    const { isLoaded, error, lastUpdatedOn, stale, counties } = this.state;
+    const { error, lastUpdatedOn, stale, counties} = this.state.parsedData;
     const cloneOfCounties = [...counties].sort(
       (a, b) => parseInt(b.value, 10) - parseInt(a.value, 10)
     );
     return (
       <Card
         error={error}
-        isLoaded={isLoaded}
         title="Incidenta cumulata a cazurilor la nivel județean în ultimele 14 zile la mia de locuitori"
         subtitle={`Ultima actualizare: ${formatDate(lastUpdatedOn)}`}
         embedPath={EMBED_COUNTIES_TABLE}
         isStale={stale}
       >
-        <div className="perDayTable">
-          <table>
+        <div className={styles.per_day_table}>
+          <table className={styles.table}>
             <thead>
-              <tr>
+              <tr className={styles.tr}>
                 <th>Județ</th>
                 <th className="has-text-right">Cazuri totale</th>
                 <th className="has-text-right">Incidență cumulată</th>
