+++ components/cards/summary/summary-card.module.css
@@ -3,6 +3,6 @@
     font-size: 0.95em;
 }
 
-.mini-chart {
+.mini_chart {
     height: 90px;
 }
