+++ components/layout/embed-button/embed-button.module.css
@@ -1,20 +1,11 @@
-.fab-action {
+.fab_action {
     position: absolute;
     top: 10px;
     right: 10px;
     border-radius: 4px;
     background-color: #f0f0f0;
     cursor: pointer;
-}
-
-.fab-action .tooltip {
     display: flex;
     align-items: center;
-    height: 30px;
-}
-
-.fab-action .fab-icon {
-    width: 16px;
-    height: 16px;
-    margin: 4px 8px;
+    padding: 6px 8px;
 }
