+++ components/layout/page-header/page-header.module.css
@@ -2,14 +2,13 @@
     margin-top: 30px;
 }
 
-.page_header h1 {
+.title {
     background: url(/images/icon-circle.svg) no-repeat;
     background-size: 1em;
     color: var(--accent-color);
     padding-left: 1.2em;
+    color: #363636;
+    font-size: 2rem;
     font-weight: 500;
-}
-
-.page_header .subtitle {
-    font-size: 1em;
+    line-height: 1.125;
 }
