--- components/layout/toast/context.jsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const ToastContext = React.createContext(null);
