--- components/loader/index.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import styles from './loader.module.css';
-
-const Loader = () => <div className={styles.loading}><div /><div /><div /><div /></div>
-export default Loader
