--- components/loader/loader.module.css
@@ -1,55 +1,0 @@
-.loading {
-    display: inline-block;
-    position: relative;
-    width: 5em;
-    height: 5em;
-}
-.loading div {
-    position: absolute;
-    top: 2em;
-    width: 0.8em;
-    height: 0.8em;
-    border-radius: 50%;
-    background: #b2bece;
-    animation-timing-function: cubic-bezier(0, 1, 1, 0);
-}
-.loading div:nth-child(1) {
-    left: 0.5em;
-    animation: loading1 0.6s infinite;
-}
-.loading div:nth-child(2) {
-    left: 0.5em;
-    animation: loading2 0.6s infinite;
-}
-.loading div:nth-child(3) {
-    left: 2em;
-    animation: loading2 0.6s infinite;
-}
-.loading div:nth-child(4) {
-    left: 3.5em;
-    animation: loading3 0.6s infinite;
-}
-@keyframes loading1 {
-    0% {
-        transform: scale(0);
-    }
-    100% {
-        transform: scale(1);
-    }
-}
-@keyframes loading3 {
-    0% {
-        transform: scale(1);
-    }
-    100% {
-        transform: scale(0);
-    }
-}
-@keyframes loading2 {
-    0% {
-        transform: translate(0, 0);
-    }
-    100% {
-        transform: translate(24px, 0);
-    }
-}
