--- pages/about.module.css
@@ -1,6 +1,0 @@
-.logo_partnership {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    padding: 30px;
-  }
