+++ pages/index.module.css
@@ -1,7 +1,3 @@
-.cards-row {
+.cards_row {
     margin-bottom: 15px;
-}
-
-.cards-row .card {
-    height: 100%;
 }
