--- public/index.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#f5bcd6" />
-    <meta
-      name="description"
-      content="Datelazi.ro prezintă infografii actualizate periodic cu datele furnizate de autoritățile competente."
-    />
-    <title>COVID-19: Date La Zi</title>
-
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="%PUBLIC_URL%/images/favicon/apple-touch-icon.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="96x96"
-      href="%PUBLIC_URL%/images/favicon/favicon-96x96.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="32x32"
-      href="%PUBLIC_URL%/images/favicon/favicon-32x32.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="16x16"
-      href="%PUBLIC_URL%/images/favicon/favicon-16x16.png"
-    />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="shortcut icon" href="%PUBLIC_URL%/images/favicon/favicon.ico" />
-    <link rel="icon" href="%PUBLIC_URL%/images/favicon/favicon.ico" />
-
-    <meta
-      property="og:image"
-      content="https://datelazi.ro/datelazi_banner.jpg"
-    />
-    <meta property="og:type" content="website" />
-    <meta property="og:url" content="https://datelazi.ro" />
-    <meta property="og:site_name" content="COVID-19: Date La Zi" />
-    <meta property="og:title" content="Date la zi" />
-    <meta
-      property="og:description"
-      content="Datelazi.ro prezintă infografii actualizate periodic cu datele furnizate de autoritățile competente."
-    />
-  </head>
-
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-  </body>
-</html>
