+++ styles/index.scss
@@ -48,16 +48,44 @@ code {
   margin-right: 10px;
 }
 
-.navigation-chevron {
-  width: 24px;
-  height: 24px;
+
+.embed_container {
+  padding: 3px;
+  overflow: hidden;
+}
+.embed_footer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.card .navigation {
-  padding-top: 5px;
-  bottom: 0px;
-  display: block;
-  width: 100%;
-  min-width: 100%;
-  height: 35px;
+.embed_footer__logo {
+  line-height: 1;
+}
+
+@media only screen and (max-width: 760px) {
+  .embed_footer .caption {
+      display: none;
+  }
+
+  .embed_footer .developed-by {
+      flex-direction: row;
+      padding: 0;
+  }
+
+  .embed_footer .developed-by .logo-wrapper {
+      padding-right: 20px;
+  }
+
+  .embed_footer .developed-by > div {
+      border: 0;
+      justify-content: space-around;
+  }
+}
+
+@media only screen and (max-width: 400px) {
+  .embed_footer {
+      padding-top: 10px;
+      flex-direction: column;
+  }
 }
